Revert "Ajustando arquitetura hexagonal"
This reverts commit 3d5779e2d1121e7420c35f41413eb7b2b22c08a4.
