--- src/domain/cliente/cliente.model.ts
@@ -1,14 +1,0 @@
-import ICliente from './iCliente';
-import Conta from '../conta/conta.model'; 
-class Cliente implements ICliente {
-    constructor(
-        public nome: string,
-        public id: string,
-        public endereco: string,
-        public telefone: string,
-        public email: string,
-        public contas: Conta[] = []
-    ) {}
-}
-
-export default Cliente;
