--- src/domain/conta-corrente/conta-corrente.model.ts
@@ -1,17 +1,0 @@
-import Conta from '../conta/conta.model';
-import tipoContaEnum from '../../Enums/tipoContaEnum';
-
-class ContaCorrente extends Conta {
-    public limiteChequeEspecial: number;
-
-    constructor(
-        codigo: number,
-        limiteChequeEspecial: number,
-        saldo: number = 0
-    ) {
-        super(codigo, saldo, tipoContaEnum.CORRENTE);
-        this.limiteChequeEspecial = limiteChequeEspecial;
-    }
-}
-
-export default ContaCorrente;
