--- src/domain/conta-corrente/conta-corrente.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ContaCorrenteService } from './conta-corrente.service';
-import { ContaCorrenteController } from '../../application/conta-corrente/conta-corrente.controller';
-import { ContaRepository } from '../../adapters/conta/contaRepository';
-
-@Module({
-  providers: [ContaCorrenteService, ContaRepository],
-  controllers: [ContaCorrenteController]
-})
-export class ContaCorrenteModule {}
