+++ src/interfaces/saqueResult.ts
